+++ examples/http_client_usage/simple_example.py
@@ -18,8 +18,10 @@ rpc_client = pyjsonrpc.HttpClient(url = "http://localhost:8080")
 print rpc_client.call("add", 1, 2)
 
 # Example with direct *call*
-print rpc_client("add", 1, 2)
+print rpc_client("add", 3, 4)
 
 # Example with *attribute* name
-print rpc_client.add(1, 2)
+print rpc_client.add(5, 6)
 
+# Example with *notify* (no response)
+rpc_client.notify("add", 7, 8)